backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-cloud-firestore==2.21.0
pydantic==2.11.7
python-multipart==0.0.12
slowapi==0.1.9
redis==5.0.1
pytest==8.0.0
pytest-asyncio==0.23.0
httpx==0.27.0
python-dotenv
firebase-admin
google-auth
openai